move provider::duration::digital to provider::duration
