--- components/experimental/src/duration/provider/mod.rs
@@ -1,10 +1,0 @@
-// This file is part of ICU4X. For terms of use, please see the file
-// called LICENSE at the top level of the ICU4X source tree
-// (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
-
-//! Data provider struct definitions for this ICU4X component.
-//!
-//! Read more about data providers: [`icu_provider`]
-
-pub mod digital;
-pub use digital::{DigitalDurationDataV1, DigitalDurationDataV1Marker};
