+++ provider/data/experimental/stubdata/digital_duration_data_v1_marker.rs.data
@@ -9,16 +9,16 @@ macro_rules! __impl_digital_duration_data_v1_marker {
         #[clippy::msrv = "1.70"]
         const _: () = <$provider>::MUST_USE_MAKE_PROVIDER_MACRO;
         #[clippy::msrv = "1.70"]
-        impl icu_provider::DataProvider<icu::experimental::duration::provider::digital::DigitalDurationDataV1Marker> for $provider {
-            fn load(&self, req: icu_provider::DataRequest) -> Result<icu_provider::DataResponse<icu::experimental::duration::provider::digital::DigitalDurationDataV1Marker>, icu_provider::DataError> {
-                Err(icu_provider::DataErrorKind::IdentifierNotFound.with_req(<icu::experimental::duration::provider::digital::DigitalDurationDataV1Marker as icu_provider::DataMarker>::INFO, req))
+        impl icu_provider::DataProvider<icu::experimental::duration::provider::DigitalDurationDataV1Marker> for $provider {
+            fn load(&self, req: icu_provider::DataRequest) -> Result<icu_provider::DataResponse<icu::experimental::duration::provider::DigitalDurationDataV1Marker>, icu_provider::DataError> {
+                Err(icu_provider::DataErrorKind::IdentifierNotFound.with_req(<icu::experimental::duration::provider::DigitalDurationDataV1Marker as icu_provider::DataMarker>::INFO, req))
             }
         }
     };
     ($ provider : ty , ITER) => {
         __impl_digital_duration_data_v1_marker!($provider);
         #[clippy::msrv = "1.70"]
-        impl icu_provider::IterableDataProvider<icu::experimental::duration::provider::digital::DigitalDurationDataV1Marker> for $provider {
+        impl icu_provider::IterableDataProvider<icu::experimental::duration::provider::DigitalDurationDataV1Marker> for $provider {
             fn iter_ids(&self) -> Result<std::collections::BTreeSet<icu_provider::DataIdentifierCow<'static>>, icu_provider::DataError> {
                 Ok(Default::default())
             }
